% Generated by roxygen2 (4.0.2): do not edit by hand
\name{colorblind_friendly}
\alias{colorblind_friendly}
\alias{colorblind_friendly_grey}
\title{Return a colorblind-friendly color palette}
\usage{
colorblind_friendly()

colorblind_friendly_grey()
}
\description{
\code{colorblind_friendly} returns a list of colorblind-friendly colors
represented as RGB hex triplets

\code{colorblind_friendly_grey} returns a list of colorblind-friendly colors
represented as RGB hex triplets. These colors are the same as those returned
by \code{colorblind_friendly}, but with grey used instead of black.
}
\note{
Source: \url{http://jfly.iam.u-tokyo.ac.jp/color/image/pallete.jpg}
}

