% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ggsave_golden}
\alias{ggsave_golden}
\title{Save a plot with proportions equal to the golden ratio}
\usage{
ggsave_golden(filename, ...)
}
\description{
\code{ggsave_golden} uses \code{\link{ggsave}} to save a plot whose
proportions follow the golden ratio (wide).
}

