% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rescale_golden}
\alias{rescale_golden}
\title{Rescale a ggplot2 plot so that axes follow golden ratio}
\usage{
rescale_golden(plot)
}
\arguments{
\item{plot}{A ggplot2 plot object}
}
\description{
\code{rescale_golden} rescales the given ggplot2 so that the axes follow the
golden ratio (horizontally).
}

