% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/straincolors.R
\name{straincolors}
\alias{straincolors}
\title{Get colors used to represent bacterial strains}
\usage{
straincolors()
}
\description{
\code{straincolors} returns a list containing strains and their corresponding
colors as RGB hex triplets.
}

