% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tableau10.R
\name{colors_tableau10}
\alias{colors_tableau10}
\alias{colors_tableau10_light}
\alias{colors_tableau10_medium}
\title{Get Tableau's 10-color palette}
\usage{
colors_tableau10()

colors_tableau10_light()

colors_tableau10_medium()
}
\description{
\code{colors_tableau10} returns a vector of RGB hex triplets representing
Tableau's 10-color color palette.

\code{colors_tableau10_light} returns a vector of RGB hex triplets
representing Tableau's lightly saturated 10-color color palette.

\code{colors_tableau10_light} returns a vector of RGB hex triplets
representing Tableau's medium saturated 10-color color palette.
}
\note{
These values come from: \url{http://tableaufriction.blogspot.ro/2012/11/finally-you-can-use-tableau-data-colors.html}
}

