% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/what_varies.R
\name{what_varies}
\alias{what_varies}
\title{Find columns that contain variation}
\usage{
what_varies(d)
}
\arguments{
\item{d}{A data frame}
}
\value{
A vector containing the names of columns that contain variation
}
\description{
\code{what_varies} finds the columns in a given data frame that contain
variation.
}
\examples{
what_varies(datasets::mtcars)
}
\references{
\url{https://gist.github.com/briandconnelly/f49c0e4d7ae11ed62fb5}
}

