% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ggsave_golden}
\alias{ggsave_golden}
\title{Save a plot with proportions equal to the golden ratio}
\usage{
ggsave_golden(filename, ...)
}
\description{
\code{ggsave_golden} is a wrapper for \code{\link{ggsave}} that saves a
ggplot to a file with proportions equal to the golden ratio (wide).
}

